- Renamed composable.adjustElementPosition to composable.autoPosition
