+++ src/ui.form-date-picker/composables/attrs.composable.js
@@ -3,7 +3,7 @@ import { cva } from "../../service.ui";
 import { computed, watchEffect } from "vue";
 
 import defaultConfig from "../configs/default.config";
-import { POSITION } from "../../composable.adjustElementPosition";
+import { POSITION } from "../../composable.autoPosition";
 
 export default function useAttrs(props, { isShownCalendar, isTop, isRight }) {
   const { config, getAttrs, isSystemKey } = useUI(defaultConfig, () => props.config);
